api_django/requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers==4.1.0
reportlab==4.4.2
django-filter==24.3
python-decouple==3.8
drf-spectacular==0.27.0
boto3==1.35.86
django-storages==1.14.4

# Testing dependencies
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
factory-boy>=3.2.0
faker>=18.0.0
model-bakery>=1.9.0

# Code quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0